Rewrite Doxie attribute retrieval logic
Newer Doxie firmwares provide more information on the 'hello' API call,
and return nothing on 'hello_extra'. There are also some newer
properties, and some missing ones depending on the hardware model.

The new logic will attempt to load attributes from 'hello' first,
falling back to 'hello_extra' if not found. If an attribute is
specified, we raise an exception as normal behavior.

For attributes we know the name of, a mapping has been provided where we
wish to translate from camelCase to snake_case.

As part of this change, all authentication handling has been moved to
the __init__ method.