Implement request retry logic for all methods
Sometimes Doxie's API will fail randomly, especially when a password is
set. Previously scan deletions retried using a for loop.

Now we implement a requests session with an adapter to retry on all
instances of 401 or 403 error codes. If we're unable to succeed within
the retry count, we let the methods handle errors as they normally
would.