Add Dockerfile
This Dockerfile builds an image on top of Python's Alpine base image. We
create a user for unprivileged execution, install doxieapi and its
dependencies, and offer the main method as the entrypoint.

Two volumes can be provided:
- /config (which should contain a doxieapi.ini for passwords)
- /scans (where scans will be downloaded to)

Auto-discovery can work if your network is bridged to allow SSDP. In
most cases this image will be run with servers defined using the
--server argument.